requirements.txt
Django>=4.0,<5.0
psycopg2>=2.9,<3.0
django-tailwind>=2.0,<3.0
djangorestframework>=3.12,<4.0
drf-yasg>=1.20,<2.0
pytest>=6.0,<7.0
pytest-django>=4.0,<5.0
